Restyle Bool constructors, move compatibility deprecations into compatibility package object
